strips out the field filters container component
