+++ src/components/CohortBuilder/Queries/index.js
@@ -1,10 +1,12 @@
 import React from 'react';
 import styled from 'react-emotion';
+import { omit } from 'lodash';
 import { memoize } from 'lodash';
 import AdvancedSqonBuilder from '@arranger/components/dist/AdvancedSqonBuilder';
 import ExtendedMappingProvider from '@arranger/components/dist/utils/ExtendedMappingProvider';
 import { withApi } from 'services/api';
 import { arrangerProjectId } from 'common/injectGlobals';
+import { FieldFilterContainer } from '../common';
 
 const extendedMappingToDisplayNameMap = memoize(extendedMapping =>
   extendedMapping.reduce((acc, { field, displayName }) => {
@@ -22,7 +24,7 @@ const SqonBuilderContainer = styled('div')`
 /**
  * this component should mimic the AdvancedSqonBuilder's API directly
  **/
-export default withApi(({ api, ...rest }) => {
+const SqonBuilder = withApi(({ api, ...rest }) => {
   const ARRANGER_API_PARTICIPANT_INDEX_NAME = 'participant';
   const onSqonRemoveClick = ({ indexToRemove, dependentIndices }) => {
     if (dependentIndices.length) {
@@ -50,6 +52,9 @@ export default withApi(({ api, ...rest }) => {
               arrangerProjectId={arrangerProjectId}
               arrangerProjectIndex={ARRANGER_API_PARTICIPANT_INDEX_NAME}
               getSqonDeleteConfirmation={onSqonRemoveClick}
+              FieldOpModifierContainer={props => (
+                <FieldFilterContainer showHeader={false} {...props} />
+              )}
               fieldDisplayNameMap={extendedMappingToDisplayNameMap(extendedMapping)}
               {...rest}
             />
@@ -59,3 +64,13 @@ export default withApi(({ api, ...rest }) => {
     </SqonBuilderContainer>
   );
 });
+
+SqonBuilder.propTypes = omit(AdvancedSqonBuilder.propTypes, [
+  'api',
+  'arrangerProjectId',
+  'arrangerProjectIndex',
+  'getSqonDeleteConfirmation',
+  'fieldDisplayNameMap',
+]);
+
+export default SqonBuilder;
